refactoring changes, added new Pin class and clockPinManager Class
